requirements.txt
pandas
numpy
requests
pymongo
python-dotenv
fastapi
uvicorn
scikit-learn
mlflow
apscheduler
